Fix achievement notifications and activity
PendingIntents for each notification were somehow being overwritten.
Added the notification ID as an intent extra and pass it as the second
parameter to the PendingIntent constructor, which is the supposedly unused
"requestCode" parameter. Set PendingIntent to FLAG_CANCEL_CURRENT, which
appears to have fixed the issue with showing the same activity again after
pressing the back button.
See stackoverflow.com/questions/3730258/mulitple-instances-of-pending-intent